requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
detectron2>=0.6

# Computer vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
pillow>=10.0.0
albumentations>=1.3.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Model export
onnx>=1.14.0
onnxruntime>=1.16.0

# Metrics
scikit-learn>=1.3.0
pycocotools>=2.0.7

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
requests>=2.31.0

# Deployment (optional)
# tensorrt>=8.6.0  # For Jetson - install separately
# pycuda>=2022.2   # For Jetson - install separately

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development
jupyter>=1.0.0
ipython>=8.14.0
black>=23.7.0
flake8>=6.1.0
